\name{viewEEG}
\alias{viewEEG}

\title{Draw up Manipulative Plot of EEG family}
\description{
This function provide manipulative plot of the Extended Exp-G family (Normal and Weibull) distribution.
With the help of sliders, you can draw different plot of density and hazard function.
You need the \pkg{manipulate} package and run function in \pkg{RStudio}
because the \pkg{manipulate} package must be run from within \pkg{RStudio}.
}

\usage{
viewEEG()
}
\arguments{
This function has no parameter.
  }

\value{
This function has no return value.
 }
\references{
Alizadeh, Morad, Mahmoud Afshari, Bistoon Hosseini, and Thiago G. Ramires.
"Extended exp-G family of distributions: Properties, applications and simulation."
Communications in Statistics-Simulation and Computation (2018): 1-16.
}
\author{Bistoon Hosseini, Mahmoud Afshari}

\examples{
viewEEG()
}

