\name{viewGOGaG}
\alias{viewGOGaG}

\title{Draw up Manipulative Plot of GOGaG family}
\description{
This function provide manipulative plot of the Generalized Odd Gamma Uniform/Weibull distribution.
With the help of sliders, you can draw different plot of density and hazard function.
You need the \pkg{manipulate} package and run function in \pkg{RStudio}
because the \pkg{manipulate} package must be run from within \pkg{RStudio}.
}

\usage{
viewGOGaG()
}
\arguments{
This function has no parameter.
  }

\value{
This function has no return value.
 }
\references{
Bistoon Hosseini, Mahmoud Afshari, and Morad Alizadeh. "The Generalized Odd Gamma-G Family of Distributions:
Properties and Applications." Austrian Journal of Statistics 47.2 (2018): 69-89.

}
\author{Bistoon Hosseini, Mahmoud Afshari}

\examples{
viewGOGaG()
}

